set up autoloader, load the classes, and rename the files
