+++ classes/class-object-sync-sf-salesforce.php
@@ -1,16 +1,15 @@
 <?php
 /**
- * Class file for the Object_Sync_Sf_Salesforce class.
+ * Authorize and communicate with the Salesforce REST API. This class can make read and write calls to Salesforce, and also cache the responses in WordPress.
  *
- * @file
+ * @class   Object_Sync_Sf_Salesforce
+ * @package Object_Sync_Salesforce
  */
 
-if ( ! class_exists( 'Object_Sync_Salesforce' ) ) {
-	die();
-}
+defined( 'ABSPATH' ) || exit;
 
 /**
- * Ability to authorize and communicate with the Salesforce REST API. This class can make read and write calls to Salesforce, and also cache the responses in WordPress.
+ * Object_Sync_Sf_Salesforce class.
  */
 class Object_Sync_Sf_Salesforce {
